Draft EVE Online bot framework
This is a rough draft for a framework to support the development of EVE Online bots. It is based on an example bot which is a warp to 0km autopilot. In this simplified version, some things are implicit which likely need to be coded on the bot side later. An example of these constraints is: How do we model restarting the client or selecting the right client instance in the bot?  The idea is to support this simplified version in the first iteration and then add the features for advanced scenarios later.