Make it easier to start bots
Support loading a bot directly from a Github URL.
Expand the set of sources to use with the `--bot-source` parameter to include Github URLs. Automate tests of this new functionality.
Provide a consistent user interface for local and web bot sources: Unify the processing of the loaded bot code to provide consistent messages for the user in case of problems in the bot source path or the loaded bot code.