Refine bot interface
Refine the interface between bot and host to integrate what I learned in the last seven weeks. This refinement is also informed by the following two observations:
+ While working on the Sanderling framework for EVE Online bots, I saw this kind of interface already evolving (See the `statusMessage` record field in `SimpleSanderling.elm`)
+ More recently, I got confused about a bot status message displayed in the engine (Also see https://github.com/Viir/bots/issues/1 for the original report). In that case, the bot did not set the status message in each step, so the engine continued to display the message from an earlier bot step. I first thought this message came out of the processing of the last event (for which the time is displayed). But with the previous implementation, the engine kept the previous status message if the bot did not respond with a `SetStatusMessage` request.

The change in this commit adapts the bot interface to integrate these learnings. The new interface requires the bot developer to explicitly state the status message, so this can not be forgotten. At the same time, this interface avoids the confusion about cases where a bot would respond with multiple `SetStatusMessage` requests in the same step.

This commit also updates all example bots to the new interface.

Since we already kept the previous bot interface for seven weeks, I expect this version to be sufficient for months to come.