Learn which use-cases and bots are more important
Also, add support for notifications and viewing the bot status on other devices. Offer the user to view the bot status from other devices such as smartphones: Report the started bot session and the status message to the botengine reactor. We can probably derive notifications from the status message, so maybe this version of the console is already sufficient to enable the notifications feature.