Improve the readability of examples and templates
Use the `Bot.elm` filename as the starting point for bot developers to customize or develop a bot. This change also means we change the convention of which file from parse the bot tags and description to use on the catalog.
Also, simplify the name of the module for the interface to the host.