Support development of EVE Online mission bot
Coordinate training data for UI elements in EVE Online.
To detect if undocking is in progress, we could check the length of the text on the button. I saw `Undocking` and `Abort Undock` (mouse hover) texts on this button after undocking was started. Another way would be looking at the saturation of the button's background: I saw the saturation going down (blue component increasing to 24 out of 255) (background color 0x4D3E18) when undocking was started.