Explore ways to improve image reading
Explore a way to avoid the stack overflow observed in decoding with larger BMP image files used in this tool. This variant implements the binning on the side of the volatile host. Also, this variant implements the decoding from file formats to pixels on the volatile host side. As a result of using imagesharp here, we get decoding of JPEG, PNG and others.