Improve the image reading framework
Change the image file decoding to support larger images.
When using the previous version on a BMP image generated from `2019-06-26.eve-online-screenshot.jpeg`, it failed with this error:
> Stopped with result: Failed to decode image: Failed to decode pixel array

This problem correlated with the size of the image; when the image has too many pixels, the bytes decoding part (in the previous version) ran into a stack overflow. `Bytes.Decode` in turn has the problem that it hides the exception by decoding to `Nothing` instead. This problem in `Bytes.Decode` has also been reported at https://github.com/elm/bytes/issues/9