Implement MutantWizards EVE Online mining bot
Implement the EVE Online mining bot described by MutantWizard.
This new bot is based on a combination of the travel bot from `eve-online-warp-to-0-autopilot` and the sanderling framework from the directory `eve-online-sanderling-framework`.
During the development, a simplified intermediate language to model the bot program emerged. The `Bot.elm` module uses this language to model the bot, composed of multiple sequences of steps. Conditional jumps (using the `JumpToSequence` instruction) support modeling branching. A new memory reading from the game client is taken for the execution of each step. In each step, the decision on how to proceed is made solely based on the current impression of the environment. This impression of the bots environment is modeled in the `MemoryMeasurement`.
Control flow is modeled using the `BotProgramStepResult` type. The `Wait` case means the same step will be executed again after a taking the next memory reading from the game client.