Introduce a simple mining bot for EVE Online
Add the first version of a simple mining bot for EVE Online based on the new development framework introduced in May. This mining bot implementation is based on the recent learnings during iterations on Mutant Wizards mining bot, as discussed at https://forum.botengine.org/t/how-to-automate-mining-asteroids-in-eve-online/628/107?u=viir

The implementation added here is based on the bot in the directory `mutant-wizard-mining-bot`. The interesting difference is in the `Bot.elm` file.

In contrast to the earlier design, this design better conforms with what we recently observed about how users think about the bot:
As we can read in the discussion thread, information about the steps were often missing in users' posts, both in observations and suggestions for changes in the bot. The state of the game was more likely to be described.
As a consequence, this design does away with the sequence-of-steps approach and instead lets the bot look at the current state of the game world to make decisions.

I illustrated this approach at https://forum.botengine.org/t/how-to-automate-mining-asteroids-in-eve-online/628/109?u=viir:

> We start with a decision tree that we use at the current impression of the environment, that is, the reading from the game client. The leaves of this decision tree are those sequences of actions. When we have completed or failed such a sequence, we go back to the root of the decision tree.
> [...]
> While I am at illustrating the architecture: Another thing that might not be obvious and should be part of the guide for developers: All branches in the tree are evaluated on the same reading from the game client. For each of the steps in the sequences the bot takes a new reading from the game client when transitioning from one step to the next.