Help developers compose bots
Bot developers focus is on the `Bot.elm` file for most changes. Most changes in bots code happen in the `Bot.elm` module. This means we can provide a way to better filter bots and changes by providing a separate ID for the part of the code which is everything else than this file.
This change introduces the bot framework ID, to identify the part of the bot code except the `Bot.elm` module. The derivation of the ID from the file tree works the same as for the bot ID. The difference is removing the `Bot.elm` file before deriving the ID.