Demonstrate how to play audio from a WAV file
Use the `--bot-configuration` parameter to specify the path to the WAV file.

A few notes about alternatives and the future:

Long-term, the easiest way will probably not require any change to the bot code but adding a passive layer. That architecture is simpler and possible to use because no information needs to flow back to the bot.

I looked into using the Notifications API in web browsers. But I did not yet find a way to configure a sound effect in chrome or firefox, so I did not further explore that route today. The approach demonstrated in this commit is an alternative quick temporary solution.