Improve readability in the repository
Remove obsolete training data for EVE Online and update the directory containing the automated tests for parsing EVE Online memory readings. Remove parts of the training data that were specific to the representation of the memory reading used with 32-bit clients. These are obsolete since the switch to memory reading for 64-bit clients.