Implement suggested functionality
Let the bot finish the session when there is nothing left to do: Finish the session three minutes after the last activity was required. This condition also covers the case where the route led to docking in a station.

See the feedback at https://github.com/Arcitectus/Sanderling/pull/10#issue-99619466