Improve robustness in case TW2 client crashes
Avoid the bot getting stuck in case the Tribal Wars 2 client stops: Reload the web page.
This change also improves the web browser framework: The previous version could cause discarding a request in a bot's response to an event in case the framework was still waiting for a task to complete. To avoid having the bot repeat redundant requests, it learns to wait while a request is pending.