Learn how to make it easier to run bots and other apps
Provide a framework for running apps: Using a batch script, we can automate some of the steps. We can automate checking if the botengine command is registered and redirect to an installation guide if necessary. Questions showed some people having problems composing the `run-bot` command or executing the command in the Command Prompt app. The script file helps people used to double-click on an icon to start an app. The packaging into a script file means we can quickly implement a graphical user interface to configure the app: The output is just an adapted script file.