Show how to add a default for the station name
Supporting the scenario described by medede at https://forum.botengine.org/t/help-with-mining-bot/3156

----
Hi I’ve been using the mining bot on my main computer and I have been having trouble getting it to return back to the station I want it to, when I start the bot in the asteroid field.
I’ve been attempting to modify the code so that there is a default value for “lastDockedStationNameFromInfoPanel” so I can start the bot while already mining and have it return to the station on its own.