Demonstrate Jetcan Mining
As described by ragnarok at https://forum.botengine.org/t/eve-jetcan-mining/3179

(Copied from the forum thread) Below is the idea following the screengrab and explanation from ragnarok:
----
I am looking at the screengrab:

What I see there looks like we should find it in the contextMenus field. I see entries containing the texts Select All and Jettison. Based on the looks in the screenshot, I expect these to end up in the entries of the ContextMenu.

Working with context menus is very common; many bots do this. Since this is such a well-beaten path, we can skip the step of using a full memory sample of the game client. The screengrab seems sufficient for the identification of the elements we want to interact with.

So I can go directly to the programming of the bot, starting with the version linked above.

Here is the part we want to replace: https://github.com/Viir/bots/blob/6923ed2ed2bed33bf07b2ea5481368f1b63df424/implement/applications/eve-online/eve-online-mining-bot/src/Bot.elm#L241-L250

To open the context menu, we can reuse the code that unloads the ore in the station. This code finds an item in the ore hold and then starts a drag&drop there to drag it to the item hangar. This means that implementation already gets us the UI element that we want to open the context menu on; we need to replace the drag&drop there with a right-click. For choosing the menu entry, we can copy from one of the many places where the bot uses a context menu.