Display how many times we unloaded the ore
As suggested by Murderburger at https://forum.botengine.org/t/mining-bot-wont-approach/3162/3?u=viir

To count the unloads, we look at the used capacity in the ore hold indicated by the capacity gauge in the game client's user interface. Each time the used volume transitions to zero, we increment the counter.
This approach does not depend on remembering what the bot tried to do because it only watches what happens in the game client. The counting also works if another bot or a human do the unloading. Since we don't need to depend on our actions for counting, this implementation is easy to transfer to other apps.