Avoid confuse "botengine app" and online service
Adapt to the recent emergence of the term "botengine app" applied to the apps that we run using the `run(-bot)` command in the Windows console app: Avoid confusing a "botengine app" with the central server providing the online services, henceforth called the reactor. Settle on the name "reactor" for this server. To help distinguish between the two, use the "reactor" subdomain to access the online services.