Fix jet can selection bug discovered by Foivos Saropoulos
Do not pick the first jet can we see in the overview, instead ensure we complete the warp to fleet member process first at least once. Relying on the bot memory here because in the current training data set, there is no way to distinguish between the jet can collection site and the site we enter immediately after undocking.
This change also means starting the bot at the mining site might not be supported anymore.

As Foivos Saropoulos aka Cerberus described at https://forum.botengine.org/t/eve-jetcan-collection/3231/12?u=viir:

```
If i understand the code correctly in line 415 :
DescribeBranch “I see no jetcan in the overview. Warp to fleet member.”
you check to see it there are jetcans in the overview and if not warp to member?
if that is the case we need to make an adjustment, insted of searching for the jet can when the ship undocks we need to check for the member insted, because if someone left a jetcan outside of the station that is going to be a problem (many ppl do that kind of think to lure new players)
```

This commit also removes the obsolete part to unlock targets.