Provide users with fast feedback on app-settings
Improve EVE Online example projects to resolve a bottleneck in app-specific feedback on app-settings for users: Configure the bot so that the framework can parse app-settings from the user directly. This way, the framework can forward the app-specific feedback directly. In earlier versions, the framework would get this feedback only via the event processing function in the bot, which was only integrated after completion of the startup sequence. For many apps, that startup sequence includes an automatic selection of the game client, and this also depends on a volatile host.
This removal of dependency on volatile hosts also enables immediate feedback in the app-settings editor: We can update the result message every time the user enters a new character, in less than a second.