Return to publishing single-file executable 🧪
Let's see if the improvements coming with dotnet core SDK 3.1.202 are sufficient to fix the problem that caused us to avoid the single-file bundling.
Last year, several users were blocked by issues specific to the single-file distribution, causing us to avoid it until now. The error reports showed several people got the "The application to execute does not exist" error on application startup:

+ https://github.com/Viir/bots/commit/6a4175830a7f13404b1ca19ad924b39154440c75
+ https://github.com/Viir/bots/issues/3#issuecomment-574999203
+ https://github.com/dotnet/core/issues/3830#issuecomment-555228146
+ https://forum.botengine.org/t/how-to-automate-mining-asteroids-in-eve-online/628/94?u=viir

For the recent improvement in .net core, see (also):
+ https://github.com/dotnet/core-setup/pull/9013
+ https://github.com/dotnet/core/blob/0e083d8449ab3c11c1dd7a769e21d12eaceb782e/release-notes/3.1/3.1.4/3.1.4.md
+ https://devblogs.microsoft.com/dotnet/net-core-may-2020/