Attempt to read the list of battle reports
The reading of the report list did not yet work in a live test: Somehow, the response string that should come via the callback did not end up where it should. The status message contained this part:

```
Read the list of battle reports: Failed to decode the response: Problem with the given value:

""

This is not valid JSON! JSON.parse Error: Unexpected input at position:0
```

But, in the dev tools console, I saw the effect of the `console.log` statements from the callback: The console showed the report list data and also the entry "Returned from callback".