Avoid troops dying on villages not fit for farming
Add new feature as requested by Bartholomeus and pandacat:

> I was wondering if there is a way to make sure the bot doesn’t send a farm to a specific village? Some barbarian villages of players who stopped playing still have walls, so everytime a farm is sent there some troops die.

This change adds a setting to specify a list of village coordinates to avoid when farming. Note that the total number of barbarian villages avoided by settings is shown in the status description. The status description contains a line like this:
```
Checked 155 coordinates and found 51 villages, 16 of wich are barbarian villages (2 avoided by current settings).
```

For the original discussion, see:

+ https://forum.botengine.org/t/farm-manager-tribal-wars-2-farmbot/3038/87?u=viir
+ https://forum.botengine.org/t/farm-manager-tribal-wars-2-farmbot/3038/107?u=viir