Combine with variation that has become popular
If the menu contains no entry named "stations", take the "structures" route. (https://forum.botengine.org/t/botengine-devlog/2183/48?u=viir)

With this change, we can use the same bot in scenarios where they want to dock at a structure, and there are no stations in the same solar system. If you want to dock at a structure, but there are also stations, remove the "stations" part from the bot code.