Fix bug causing not seeing own villages sometimes
Fix a problem that could sporadically cause us to see less of our own villages.
Caleb reported this problem 2020-07-03 at https://forum.botengine.org/t/farm-manager-tribal-wars-2-farmbot/3038/150?u=viir
In their screenshot we can see also the two following parts of the status text:
+ "Completed first farm cycle 16 minutes ago with 1 villages"
+ "Found 43 own villages"
The bot reported finishing the last farming cycle with just one village, but at the time of the screenshot was seeing many more villages.

It looks like the `selectedCharacter.data.readyVillages` contains not always all the village IDs, and the previous program code did not take this into account. To fix this problem, we now compare the number of elements in `readyVillages` with the result from `getTotalVillages`. As long as they don't match, we reread the game root information.