Fix bug in the user interface for offline sessions
Fix bug that caused the wrong message displayed when a session ended because of the time limit for offline sessions.
The previous version displayed this wrong message:

> Session finished: Session has reached configured time limit of 15 minutes.

In a live test with this version (after this change), I got the right message:

> Session  finished: Reached maximum duration for offline session of 15 minutes. To run the app in a longer session, start an online session. For more information, continue at https://to.botengine.org/from-console-reached-maximum-duration-of-offline-session

Also, make it easier to copy text from the console window when the session has finished: (Re)enable the "Quick Edit Mode".