Add image from ragnarok to describe their scenario
Add the image ragnarok shared at https://forum.botengine.org/t/eve-jetcan-mining/3179/2 to illustrate their jetcan mining process.

Below is the text I found in the forum post that linked the image:
----
Hi and thanks for the quick response!

Logic wise, it will follow the same rules as the current mining bot, the only difference would be once the ore hold reaches capacity instead of returning to station, it would take the following steps;

1) Right click on the ore and choose select all
2) Right click again and choose the option to Jettison
3) Continue to mine

Below is a screengrab of where the select all and jettison options are in the menu.

[https://i.gyazo.com/4d91c640a234b003d394ef106a9306e5.png](https://i.gyazo.com/4d91c640a234b003d394ef106a9306e5.png)

Thanks,

R