Adapt to discoveries of Anthony James on menu structure
As described today by Anthony James at https://forum.botengine.org/t/eve-online-bot-wont-return-to-its-undocking-location/3454/6?u=viir, the menu entry for a structure is not always directly under the "Structures" menu.
As Anthony described in the linked post, the station will be visible in the overview, but the viewport is not large enough in some cases so that scrolling is necessary.