Introduce simulations to test apps fast and easy🚤
Add the first version of the function to run an app in a simulation. Add the `simulate-run` command for starting a simulation. Support replaying a past session to define a simulation.

Compared to version 2020-07-17, this one also fixes a bug causing long delays in notifying an app about a completed task.