Display count and names of other pilots in the overview
Add the function `getNamesOfOtherPilotsInOverview` that takes a reading from the game client and returns a list of other pilots currently in the overview. To decide if an overview entry represents a pilot, we compare the name shown on the overview entry with the names shown in the local chat window.