Add the ID of the current anomaly to the status text
Add function to check if a probe scan result represents the anomaly we are currently in.
Add the ID of the current anomaly to the status text.

Why are we interested in this information in the first place?
We can use this information about the current anomaly to link it to observations in other parts of the user interface, like the overview window. This connection allows us to store observations related to an anomaly in the memory. This memory is useful when we need to decide which anomaly to enter next. For example, we might have found that an anomaly contains a dangerous NPC that we want to avoid.

How do we extract the ID of the current anomaly scan result from the reading of the game client?
With the reading from the game client, we get the list of scan results in the probe scanner. The probe scanner displays the `Distance` for each scan result. We use the distance text to decide if a scan result represents the anomaly we are currently in. The game uses the units "m" (meters), "km" (kilometers), and "AU" (astronomical units) to display the distance. If we see meters or kilometers as the unit, we assume we are currently in this anomaly.
We could also parse the distance text to a number (as is done, for example, for overview entries) to apply an even stricter constraint. But that seems not necessary so far.