Fix throughput for reading information about the game world
Adapt the app program code to restore the reading rate of about two Herz when searching the map for barbarian villages.
The combination of engine 2020-08-21 and the previous version exhibited a much slower reading rate. Some recent improvements in the engine led to faster processing of app events, and it looks like the app code made some assumptions about the spacing of events that became incorrect with the newer engine.

The changes in the code:

+ Fix the bug mentioned above: Change the branch in `pendingRequestTimeInMilliseconds` regarding the case of equality.
+ Expand the output in the decision tree descriptions to support inspection.
+ Adapt the web browser bot framework to generate less superfluous events.
+ Some formatting for readability.

The `gotTimeSinceTaskCompleted` field indicates apps can be simplified by sending the time with each app event.