Simplify app program codes for interface to host event structure
Adapt to observations in application projects: Integrate the time with every event: Publish botengine for Windows version 2020-08-24, supporting the new app event structure in the interface to the host. Engine version 2020-08-24 still supports running apps with the older event structure.

Adapt the example projects for the new app event structure.