Remember visited anomalies and display counter
+ Expand the memory to remember the visited anomalies and the names of other pilots we found in each on arrival.
+ Expand the memory to support tracking arrival on the grid: Add a field to remember if we were warping in the previous reading.
+ Expand the status text to display a counter of visited anomalies.

To remember the visited anomalies, we use the standard function to update the memory each time we get a reading from the game client. We define the event of arrival in an anomaly by the transition of the warping indication in the ship UI: If we were warping in the previous reading and not warping in the current one, then we just arrived.