Demo how to engage drones for Mactastic08
Based on https://forum.botengine.org/t/orca-targeting-mining/3591:

To write a program to automate these behaviors, I search the catalog for these functions and combine them into a new app. To find a building block to engage drones, I search for a ratting bot and find this one: https://catalog.botengine.org/0eac4011309d710f71b1e69c8a9878a331c04b338140a38aee8f0b88e0310a00
In this bots source code, I find this function to launch and engage drones:
https://github.com/Viir/bots/blob/00d6647c2dbc804db4ff7abbac85738390a7301f/implement/applications/eve-online/eve-online-combat-anomaly-bot/BotEngineApp.elm#L540-L581
In this function code, we also see the behavior to engage the drones described above.
To make a new bot, we can copy one of the existing popular bots and replace the decision tree's root. Using a decision tree to pick the next action is a typical pattern that you find in mining, ratting, autopilot, and other bots. These different kinds of bots share the same language/type to describe the outcome of a decision tree so that we can copy branches and reuse them in other apps.
We can copy this function as-is and use it at the root of the decision tree. (We don't need to worry about the branch for launching drones, because the bot will only enter it if necessary.)