Demo how to branch on effects from the previous step
Use the activation of module buttons in EVE Online as an example: Adapt example apps to wait with clicking on a module button if it already clicked on that one in the previous step.
Also, simplify the interface to the volatile host: Consolidate some redundant structures on the serial interface.