Update reading of army presets when starting a new farm cycle
Implement the feature idea Drklord shared at https://forum.botengine.org/t/farm-manager-tribal-wars-2-farmbot/3038/189?u=viir

```
Is there a way, when it finishes a break and refresh the browser the bot would check if there is a new compatible preset it can use? I like to add new preset, while the bot is running but then I would have to completely restart the bot so it can recognize the new preset.
```

How do we implement this behavior?
First of, to clarify the concepts: Finishing a break means starting a new farm cycle.
Some background on how the bot works: The bot reads the list of the army presets from the game client and then stores it to pick the right one when sending an attack.
To implement this new feature, we start by moving the memory of the army presets from the global level down to the level of the farm cycle. This means moving the `getArmyPresetsResult` field from the `BotState` type to the `FarmCycleState` type. We discard the farm cycle state when we finish a farm cycle, so moving the field into there protects us from accidentally using an outdated reading. After moving the field, we follow the compiler's error messages to adapt all places in the program that update or read this field.