Prepare framework for running multiple bots on same desktop
Prepare a framework for EVE Online to support multiple bots coordinating to share input focus on a single desktop: Expand the interface to the volatile host to support bundling a sequence of effects for the game client into one request.

Quoting the discussion at https://forum.botengine.org/t/combat-anomaly-bot/3720/2:

```
You can run multiple bots on the same machine without this level of isolation. To do this, you let the bots coordinate with each other and take turns using the input focus (mouse). You let each bot check if another bot is sending input at the moment before starting to send inputs itself. For EVE Online, 300 Milliseconds per step should be enough time in most cases. It depends on how you configure the distance between individual steps in the input sequence. For example, how long you want to wait after moving the mouse before pressing a mouse button. (You can see the actual times for a session of your choice using the time-travel functionality in devtools)
```

This change in the framework supports that coordination by bundling effects coming out of a single event response.

This change also improves consistency in the casing of names on the interface to the volatile host.

Notes on the current implementation:

+ The readability in the volatile host script can be further improved by cleaning up older stuff. Changes there should be tested in the alternate UI first, though.
+ As the console beep variant becomes obsolete, the framework code can be simplified further.
+ The framework could be expanded to support specifying the delay time between elements.