Optimize for the scenario shared by Max last week
Adapt to the experience Max shared 2020-11-28 at https://forum.botengine.org/t/farm-manager-tribal-wars-2-farmbot/3038/190?u=viir:

Adapt to scenarios with a huge amount of villages: Avoid getting stuck in a cycle by stricter conditions to switch to another village.

Quote from the linked thread:

```
I have been noticing some problems when I use the bot with an increasing amount of villages.
I`ll try explaining the problem (maybe its a problem only imo): since the fact that the bot is iterating over all the villages, checking if they all have 50 attacks out, before ending the farm cycle. When you have a huge ammount of villages, you start having to iterate millions of time on all the villages and using light cav. everytime some attacks came back in the meanwhile. Till now the behaviour of the bot seems pretty dope, the problem is that instead of attacking again the closest villages, in order to attack different barbarian, it starts attacking barbarians which are really far away. I know that logically is the right behaviour but there are some problems with this:
```