Guide users to configure the overview window in the game client
Avoid the issue described by nexty5870 at https://github.com/Viir/bots/issues/31

Draw the user's attention to this part of the game client's setup: Add a branch to ask the user to enable the 'Name' column in the overview window. Other functions in this app use the name field to classify overview entries as asteroids.