Improve multi-instance support - optimize runtime expenses
To learn more about the motivation, see the discussion at https://forum.botengine.org/t/farm-manager-tribal-wars-2-farmbot/3038/216?u=viir

Avoid the number of browser tabs in chrome growing every time the bot reloads the web page: Return to a model where we close previous instances but introduce a scoping that helps avoid interference between multiple instances of the bot:

+ Implement the scoping using different profile directories in the browser instances, this time for real: Complete the code on the volatile host side to read the `user-data-dir` from the browser process command line.
+ Adapt the flow to start a browser instance in the framework to support the app choosing the user profile ID for the browser instance.
+ Help the user configure the profile ID by adding an app-setting.
+ Expand the status text so the user can check which profile ID they configured in the running app.

The approach implemented here means: If the user wants to run multiple instances in parallel, they need to adapt app-settings to ensure they use different profile IDs. Note that things like bookmarks and cookies belong to that profile in the Chrome web browser. That means the user needs to log in to the game for each new profile they create. (Creation of the user profile in the web browser is still automatic by starting the bot, as before. But each time they change the profile ID, they lose the state mentioned before)

A nice property of this approach is that it should work the same for users only running a single instance. So this could become the recommended (main) version for all.