Demo how to count total number of targets in session
As discussed at https://forum.botengine.org/t/i-would-like-to-use-survey-scanner-for-every-even-numbers-of-targets-targetted/3807

You can increment the counter when a new target appears in the current reading from the game client. Thinking about it again, even better is not just to increment it, because of this scenario: You could have two new targets appearing at the same time.
Here is how this is usually implemented: In the function to update the memory, add the number of new targets to `timesTargeted`.
How do we know the number of new targets? By subtracting the number of targets in the current reading by the number of targets in the previous reading.
How do we know the number of targets in the previous reading? By storing it in the bot memory. That means we add another field to store that count to the memory type:

The same approach to counting is already implemented to count the amount of unloaded ore. (This is the count that you see in the overall status text) For the ore counter, the fields in the memory are `lastUsedCapacityInOreHold` and `volumeUnloadedCubicMeters`.