Fix bug - do not decrease timesTargeted
As discussed at https://forum.botengine.org/t/i-would-like-to-use-survey-scanner-for-every-even-numbers-of-targets-targetted/3807/5?u=viir

That subtraction results in a negative number when a target disappears. So we add a branch not to add the difference to the counter when it is negative.

Again, we can look at the ore volume counter to see it also has this branch. Here is a comment from the old code:

> During mining, when new ore appears in the inventory, this difference is negative.

The difference is that the ore counter used the `max 0` function to branch. I guess showing the `if` `then` `else` keywords inline makes it easier to understand.